pyproject.toml
[project]
name = "bahea-cal"
dynamic = ["version"]
description = "Bahea Calendar is a project that adds the standings of your favorite soccer team games to your google calendar!"
readme = "README.md"
requires-python = ">3.10"
license = {text = "MIT"}
keywords = []
authors = [
    {name = "Fernando Mota", email = "nando.mot@gmail.com"},
    {name = "Ivan Neto", email = "ivan.cr.neto@gmail.com"},
    {name = "Kauan Nery", email = "kauannery0500@gmail.com"},
    {name = "Ruan Claudio", email = "ruanclaudio4@gmail.com"}, 
    
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "oauth2client>=4.1.3",
    "google-api-python-client>=2.127.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "attrs>=23.2.0",
    "beautifulsoup4>=4.12.3",
    "arrow>=1.3.0",
    "requests>=2.31.0",
    "Django>=5.0.4",
    "dynaconf>=3.2.5",
    "djangorestframework>=3.15.1",
    "sentry-sdk>=2.3.1",
    "django-model-utils>=4.5.1",
    "boto3>=1.34.116",
]
version = "0.1.0"

[project.urls]
Documentation = "https://github.com/ivancrneto/bahea-cal#readme"
Issues = "https://github.com/ivancrneto/bahea-cal/issues"
Source = "https://github.com/ivancrneto/bahea-cal"

[project.optional-dependencies]
prod-only = [
    "gunicorn>=22.0.0",
    "gevent>=24.2.1",
    "psycopg2-binary>=2.9.9",
]

[tool.coverage.run]
source_pkgs = ["bahea_cal", "tests"]
branch = true
parallel = true
omit = [
  "bahea_cal/__about__.py",
]

[tool.coverage.paths]
bahea_cal = ["bahea_cal", "*/bahea-cal/bahea_cal"]
tests = ["tests", "*/bahea-cal/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pdm]
distribution = false
[tool.pdm.scripts]
migrate = "python manage.py migrate"
start = "python manage.py runserver"
post_init = { composite = [
    "pdm install",
    "migrate",
    "start",
] }

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.2.0",
    "coverage[toml]>=7.5.0",
    "ipdb>=0.13.13",
    "pytest-django>=4.8.0",
    "black>=24.4.2",
]

[tool.pytest.ini_options]
DJANGO_CONFIGURATION = "webapp.settings"
python_files = ["tests.py", "test_.py", "_test.py", "testing/python/*.py"]

[tool.black]
line-length = 120
target-version = [ "py38" ]
exclude = '''
(
  /(
      \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
| migrations
)
'''

[tool.isort]
atomic = true
case_sensitive = true
filter_files = true
import_heading_firstparty = "Internal imports"
import_heading_stdlib = "Python imports"
import_heading_thirdparty = "Pip imports"
known_startup = [
  "startup",
]
line_length = 120
lines_after_imports = 2
profile = "black"
py_version = "auto"
remove_redundant_aliases = true
sections = [
  "FUTURE",
  "STARTUP",
  "STDLIB",
  "THIRDPARTY",
  "FIRSTPARTY",
  "LOCALFOLDER",
]

[tool.bandit.assert_used]
excludes = ['*_test.py', '*/test_*.py']